chore(workflow): break workflow-command<->run-state-command cycle; add madge pre-push gate
A `madge --circular` audit (type-only imports excluded) found exactly one
real runtime circular dependency: workflow-command.ts <-> run-state-command.ts,
both directions value imports. It worked only because the functions are
called at runtime, not module-load — latent TDZ fragility.

Fix: extract the shared CLI helpers run-state-command pulled back from the
dispatcher (parseArgsStrict, runCliPreflightLint, plus their local helpers
printDiagnostics / loadAndValidateDefinition / formatDiagnostic) into a new
leaf src/workflow/cli-shared.ts. Both command modules now import from the
leaf; run-state-command no longer imports workflow-command. Relocation only,
no behavior change.

Gate: add `madge` as a devDependency, a `.madgerc` with skipTypeImports
(mandatory — without it madge reports 20 harmless type-only cycles and the
signal is lost), a `check:cycles` npm script, and a `.hooks/pre-push` hook
(installed via the existing `setup-hooks` convention) that blocks a push
introducing a runtime cycle. `npm run check:cycles` now reports 0.
