+++ .madgerc
@@ -1,0 +1,1 @@
+{ "detectiveOptions": { "ts": { "skipTypeImports": true }, "tsx": { "skipTypeImports": true } } }
